Fix for unnecessary error log messages when symlinked file backup

+++ urbackupcommon/os_functions_lin.cpp
@@ -305,7 +305,7 @@ bool os_directory_exists(const std::wstring &path)
 	return isDirectory(path);
 }
 
-bool os_remove_nonempty_dir(const std::wstring &path, os_symlink_callback_t symlink_callback, void* userdata)
+bool os_remove_nonempty_dir(const std::wstring &path, os_symlink_callback_t symlink_callback, void* userdata, bool delete_root)
 {
 	std::string upath=Server->ConvertToUTF8(path);
 	std::vector<SFile> tmp;
@@ -410,9 +410,12 @@ bool os_remove_nonempty_dir(const std::wstring &path, os_symlink_callback_t syml
 		if(!b)
 		    ok=false;
     }
-    if(rmdir(upath.c_str())!=0)
+	if(delete_root)
 	{
-		Server->Log("Error deleting directory \""+upath+"\"", LL_ERROR);
+		if(rmdir(upath.c_str())!=0)
+		{
+			Server->Log("Error deleting directory \""+upath+"\"", LL_ERROR);
+		}
 	}
 	return ok;
 }
